app/pyproject.toml
[tool.poetry]
name = "fastapi-s"
version = "0.1.0"
description = ""
authors = ["Stepan Andreev <toranodora@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.36"
psycopg = "^3.0.14"
alembic = "^1.7.7"
passlib = {extras = ["argon2"], version = "^1.7.4"}
psycopg2 = "^2.9.3"
pydantic = "^1.9.1"
email-validator = "^1.2.1"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptograghy"], version = "^3.3.0"}
pytest-asyncio = "^0.18.3"
pytest = "^7.1.2"
httpx = "^0.23.0"
Faker = "^13.12.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
